Making the proxy pkg better to handle
